requirements.txt
pandas==2.2.2
requests==2.32.2
lightgbm==4.3.0
mlflow==2.11.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dateutil==2.9.0
